requirements.txt
asgiref==3.11.1
attrs==26.1.0
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
defusedxml==0.7.1
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
httpie==3.2.4
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.17.2
PyJWT==2.12.1
PySocks==1.7.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
